Changing Firmware error structure to use SEV errors and modifying IOCTL calls to bubble correct errors
The AMD PSP provides defined SEV errors in the API with clear codes and
mistakes. The library currently only bubbles up IO errors and not the
correct SEV error when IOCTL calls fail. This commit rearranges the
error file to use SEV errors correctly, and then changes the way IOCTL
calls returns errors, so that the appropriate SEV error is bubbled up
when an error is encountered.

Guest IOCTL calls do not fail the same way host IOCTLS do. You need to
manually check for failures even if the ioctl function itself might not
return an error.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>